+++ from_json/date_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 
 import { test, assert, assertEquals } from "../test_deps.ts";
-import { createDateStrategy, ISODateFromJson } from "./date_from_json.ts";
+import { createDateStrategy, ISODateFromJson } from "./date.ts";
 import { Serializable, SerializeProperty } from "../mod.ts";
 
 test({
