+++ to_json/default_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 Gamebridge.ai authors. All rights reserved. MIT license.
 
 import { test, assertEquals } from "../test_deps.ts";
-import { defaultToJson } from "./default_to_json.ts";
+import { defaultToJson } from "./default.ts";
 
 test({
   name: "defaultToJson passes the value unmodified",
